Separated tests for matching and readGalfast.
